requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.5.3
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
flake8==3.7.7
flake8-mypy==17.8.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Redis==0.3.0
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.15
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==6.0.0
multidict==4.5.2
mypy==0.670
mypy-extensions==0.4.1
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
numpy==1.16.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pyrsistent==0.14.11
pytest==4.3.1
pytest-cov==2.6.1
pytest-flask==0.14.0
pytest-mock==1.10.1
pytest-vcr==1.0.1
python-dateutil==2.8.0
python-dotenv==0.10.1
PyYAML==5.1
pyzmq==18.0.0
qtconsole==4.4.3
redis==3.2.1
requests==2.21.0
requests-futures==0.9.9
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
typed-ast==1.3.1
ujson==1.35
urllib3==1.24.1
vcrpy==2.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
yapf==0.26.0
yarl==1.3.0
